requirements.txt
argunparse
bert
dotmap
ffmpeg
matplotlib
ninja
Ninja
numpy
omegaconf
opencv-python
open3d
pandas
pillow
prettytable
pyglet
pyopengl==3.1.0
pyrender>=0.1.23
pytorch-lightning
PyYAML
scikit-learn
scipy
sentence-transformers
smplx
spacy
tensorboardX
trimesh>=2.37.6
tqdm

git+https://github.com/otaheri/chamfer_distance.git
git+https://github.com/otaheri/MANO.git
git+https://github.com/otaheri/bps_torch.git
git+https://github.com/MPI-IS/mesh.git
git+https://github.com/openai/CLIP.git